MGMT-9989 - Patching the Node Root CA by hand instead of using MCO, only on Spoke/Edge clusters
This is a big change, we are using this alternative method in order to avoid reboot the Spoke Cluster nodes (which takes 45 Minutes). This method should not cause any collateral issue because all the customer workloads will be added after the spoke relocation, so the node will be rebooted in the end.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>