MGMT-10240 - Mask crio-wipe systemd unit
Masking this unit allows us to avoid the image cache on the nodes once the enclosure is relocated.
The problem happens when a power cycle appears and the service crio-wipe deletes the node images stored in cache. This issue becomes a bigger issue, knowing that the cluster is disconnected and the nodes are using their own registry to boot the pods up.
This is the main reason for this improvement.

Signed-off-by: Juan Manuel Parrilla Madrid <jparrill@redhat.com>