Migrate mirror steps to use oc-mirror
This is quite a big change and it did a refactor on several areas of the
code base. The gist and most important parts of this change are:

- The ocp-mirror and olm-mirror have now been merged
- The mirror is now done using `oc-mirror`
- The mirror step only syncs the required packages from selected
channels instead of downloading the latest from every channel as it was
doing before. Architectures are not filtered.
- There is now a single organization in Quay called `ztpfw`
- ICSPs and catalogsources are created by oc-mirror
- The catalogsource names are kept as the upstream ones to avoid having
to mess with the subscriptions.

Minor changes:
- Several sleeps have been removed in favor of waiting for resources