Add basic Go skeleton
This patch adds a basic skeleton of the Go code that will be used to
progressively replace the shell scripts of the project. Initially it
only has a basic `version` command that prints some information about
the version of the project.

To run it run change to the `go` directory, run `go build -o ztp` and
then run the resulting binary:

```
$ cd go
$ go build -o ztp
$ ./ztp
Build commit: abbfda355fc5938e544990ac8c3bcd469b45bf61
Build time: 2022-11-24T14:43:10Z
```

To test it run `go test ./...`, or for better output install the
_ginkgo_ tool and then run `ginkgo ...`:.

```
$ cd go
$ go install github.com/onsi/ginkgo/v2/ginkgo@v2.5.1
$ ginkgo -r internal
```

The patch also adds a new `check-cli` GitHub workflow to build and run
the tests of the new code.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>