Use matching ginkgo version in GitHub actions
Currently the GitHub action that checks the CLI installs a version of
the `ginkgo` tool that may not match the version that the tests are
compiled with. To address that this patch changes the action so that it
checks the version in use with the `go list ...` command and installs
it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>