NO-ISSUE: Add Kubernetes API test infrastructure
This patch adds a new `testing.Environment` object that creates a
minimal Kubernetes cluster with etcd and an API server. This is used for
tests that need to connect to a Kubernetes cluster, but it doesn't have
the cost of a complete Kubernetes cluster, just enough to run unit
tests.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>