NO-ISSUE: Wait for API server ready before running tests (#658)
Currently the testing environment starts a Kubernetes API server but
doesn't wait for it to be ready. In some situations this can cause a
problem because tests may try to use it when it isn't ready yet, and
will fail. This patch changes the environment so that it waits till the
API server is ready according to the `readyz` endpoint.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>