NO-ISSUE: Check existence of called template
Currently when a template uses the `execute` function to execute
another template the code doesn't check if that template exists. The
result is a nil pointer exception that is difficult to understand. This
patch changes the code so that it checks that condition and generates an
error message that is easier to understand.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>