NO-ISSUE: Explicitly load kbueconfig in client (#673)
Currently the Kubernetes API client loads the configuration using the
`clientcmd` package. That in turn uses the `KUBECONFIG` environment
variable. That complicates testing because it is difficult to change the
values of the environment variables and then restore them after running
the test. To make that simpler this patch changes the client builder so
that it receives a map containing the environment, and uses those
variables instead of `os.LookupEnv`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>