NO-ISSUE: Replace renderer with applier
Currently the commands that create objects from templates use the
renderer object to generate the objects from templates, and then they
create those objects explicitly. To avoid repeating that in all those
commands this patch replaces the renderer with a new "applier" type
which knows how to do the rendering and also how to actually create the
objects.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>