NO-ISSUE: Take pull secret from hub
Currently the `create cluster` command takes the pull secret from the
`PULL_SECRET` environment variable because that is what the shell
scripts do. But the shell scripts also populate that environment
variable from the `pull-secret` secret of the `openshift-config`
namespace of the hub. The result is that the command may generate a
wrong pull secret for the edge clusters, because the file pointed by the
`PULL_SECRET` environment variable may be old. To avoid that this patch
changes the command so that it gets the pull secret directly from the
hub cluster.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>