NO-ISSUE: Fix calculation of DNS domain
The DNS domain is currently copied from the configuration of the
ingressc controller of the hub. For example, if that configuration is
`apps.my-hub.my-domain.com` then the base DNS domain of the edge
clusters will also be `apps.my-hub.my-domain.com`. But that is wrong; it
should be only `my-domain.com`. This patch fixes that.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>