NO-ISSUE: Load Kubeconfig fron default locations (#723)
Currently the the tool creates the Kubernetes API client loading the
`KUBECONFIG` environment variable explicitly. This is done in order to
generate better error messages, but it generates problems when running
the tool in environments where the configuration has to come from other
places, specially when it has to come from the secrets mounted inside a
Kubernetes pod. To address this issue this patch changes the tool so
that it uses the same mechanism used by tools like `kubectl`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>