NO-ISSUE: Fix naming of nodes (#730)
Currently if the name of a node in the configuration file is `master0`
the resulting node name will be `ztpfw-...-master-master0`. That works
fine for the creation of the cluster, but it breaks the scripts that run
later to create the registry because they expect the nodes to be named
`ztpfw-...-master-0`. It would be good to make that code independent of
the names assiged to nodes, but for now we will instead change the CLI
so that it generates the expected names.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>