NO-ISSUE: Improve waiting for deleted
Currently the applier waits for objects to disappear before deleting the
namespaces. But it does that by waiting for the first delete event for
the object. That doesn't work correctly if the object has finalizers, as
the API server will not inmediately delete the object. To improve that
this patch changes the applier so that it explicitly check if the object
exists after receiving the delete event.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>