NO-ISSUE: Don't explicitly delete cluster deployments (#752)
Currently the `delete cluster` command deletes all the objects that
`clreate cluster` creates. But the Hive cluster deployment can't be
deleted because then Hive automatically deletes the namespace it
contains. As a result the bare metal hosts can't be deleted successfully
because the deletion process needs to create other objects inside the
(now gone) namespace. To address that this patch changes the `delete
cluster` command so that it doesn't delete the cluster deployment; it
will be deleted automatically when the namespace is deleted.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>