NO-ISSUE: Fix wait for CRD
Currently the code that waits for CRD to be created doesn't work
correctly because it takes the group and kind from the metadata of the
objects in events it receives, which will always be
`apiextensions.k8s.io/v1` and `CustomResourceDefinition`. It should
instead use the `.spec.group` and `.spec.names.kind` fields of the
object. This patch fixes that.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>