NO-ISSUE: Improve wait for CRD
Currently the code that waits for a CRD to be created fails occasionally
because when a CRD is created and established it may not yet be
available for use: the API server will keep responding with errors when
trying to create objects of that type. In order to react better to that
this patch changes the applier so that it repeatedly tries to create the
object till the API server stops responding with the error that
indicates that the CRD is not available.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>