NO-ISSUE: Add support for custom log fields (#762)
This patch adds command line options that add custom log fields. For
example, to add a field with name `user` containing the user name:

```
$ ztp create cluster -c my-config.yaml --log-field "user=${USER}"
```

This is intended for situations where multiple processes write to the
same file. Adding these fields it is easier to identify which process is
the one that wrote a particular log message. In that context it is
specially useful to add the PID of the proccess. The special value `%p`
is used for that:

```
$ ztp create cluster -c my-config.yaml --log-field pid=%p
```

Or just `%p`, without the `pid=` prefix:

```
$ ztp create cluster -c my-config.yaml --log-field %p
```

Note that before this patch the `pid` field was always added, now it
isn't added by default.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>