NO-ISSUE: Add options to enable/disable HTTP logging
Currently when log level 2 is enabled we write to the log all the
details of HTTP requests and responses, including HTTP headers and sizes
of HTTP request and response bodies. This is quite noisy, and usually
not necessary. This patch adds two new flags `--log-headers=[true|fase]`
and `--log-bodies=[true|false]` that allow selectively
enabling/disabling those details.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>