NO-ISSUE: Replace console args
This patch changes the console so that it replaces the arguments passsed
to the `Info`, `Warn` and `Error` methods with human friendly strings.
Currently that applies only to Kubernetes objects, which are replaced by
strings containing the namespace and name. For example, a call like
this:

```
config := &corev1.ConfigMap{
	ObjectMeta: metav1.ObjectMeta{
		Namespace: "my-ns",
		Name:	   "my-name",
	},
}
console.Info("Object '%s' doesn't exist")
```

Will result in the following output:

```
I: Object 'my-ns/my-name' doesn't exist
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>