NO-ISSUE: Add `data` template function (#777)
This patch adds a new `data` template function that builds maps whith
string keys. For example:

```
{{ range $name, $value := data "X" 123 "Y" 456 }}
{{ $name }}: {{ $value }}
{{ end }}
```

This is specially useful to pass multiple named parameters to other
templates. For example, a template that two values named `Name` and
`Age` can be executed like this:

```
{{ execute "my-template.yaml" (data "Name" "Joe" "Age" 52) }}
```

If the number of arguments isn't even, or if any of the names isn't a
string an error will be returned.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>