NO-ISSUE: Add `dev ssh` command (#782)
Currently development often requires connecting via SSH to the nodes of
the clusters. Building the SSH commands required for that is tedious and
error prone. To make that simpler this patch adds a development utility
that simplfies connecting via SSH to the clusters. If there is only one
cluster in the configuration file the following command will find the
SSH key of the cluster and the IP address of the first node and connect
to it with the `core` user:

```
$ ztp dev ssh -c config.yaml
```

The `--cluster ...` and `--node ...` options allow connecting to
specific cluster when there are multiple clusters in the configuration
file or to a specific node:

```
$ ztp dev ssh -c config.yaml --cluster my-cluster --node worker0
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>