NO-ISSUE: Delete CRDs when deleting metal load balancer
This patch changes the `ztp delete metallb` so that it also deletes the
CRDs in the groups `metalb.io` and `nmstate.io` groups if the
`--crds=true` option is used. This is intended for use in development
environments where it is convenient to completely delete them to be able
to fully test the `ztp create metallb` command.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>