NO-ISSUE: Add `delete lso` command
This patch adds a new `ztp delete lso` command that deletes the objects
created by the `ztp create lso` command. This is intended for use in
development environments, where it is convenient ro un the create
command multiple times without having to manually delete the objects or
else create the cluster from scratch.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>