MGMT-13854: Fail installation if it moves to `error`
Currently the `ztp create cluster` command exists with an error code
when the `Failed` condition of the cluster changes to `True`, but it
doesn't exit if the `.status.debugInfo.state` changes to `error`. This
patch changes it so that it also exists with an error code in that case.

Related: https://issues.redhat.com/browse/MGMT-13854
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>