MGMT-13854: Restart watches
Currently the CLI uses the controller runtime `Watch` method to wait for
events, for example to wait till the cluster is installed. But those
watches may be terminated by the server at any time. In particular we
see some cases where they terminate after one hour. As a result of that
the CLI finishes assuming that the cluster is installed, when it isn't.
To address this this patch changes the `Watch` method so that it
restarts the wait when the server finishes it.

Related: https://issues.redhat.com/browse/MGMT-13854
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>