NO-ISSUE: Add `dev shell` command (#816)
This patch adds a new `dev shell` command that runs a shell with
environment variables and SSH configuration prepared to easily connect
to the cluster. For example:

```
$ ztp dev shell --config my-cluster.yaml
I: Collecting information for cluster 'edgecluster0-cluster'
I: Kubeconfig is available, you can use 'oc' or 'kubectl' to connect to the cluster
I: You can use 'ssh' to connect to nodes 'master0', 'master1', 'master2' and 'worker0'
[user@my-cluster ~]$
```

Witin that shell the `KUBECONFIG` environment variable points to the
admin kubeconfig of the cluster, and the `ssh` and `scp` commands are aliases
that allow connecting to nodes. For example:

```
[user@my-cluster ~]$ oc get pods -A
[user@my-cluster ~]$ ssh master0
[user@my-cluster ~]$ ssh master1
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>