MGMT-13924: Restart watches when API server is restarted
When the API server is restarted the existing watches are closed, and
trying to start them again may result in requesting a resource version
that is too old and has alreayd been forgotten/compacted by etcd. That
results in receiving an `ERROR` event with code `420 Gone` indicating
that the resource version is too old. In that situation the CLI
currently just tries to create the watch again, but using the same
resource version. That results into an infinite loop. This patch fixes
that issue. When the CLI receives the `ERROR` event it will restart the
watch from version `0`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>