NO-ISSUE: Fix waiting for metal load balancer
The `create metallb` command waits till the API is reachable via metal
load balancer, but it does this reusing a client that it creates outside
of the retry loop. This client will fail to create the first time
because it sends requests to get the API metadata that will fail. To fix
that issue this patch changes that command so that it creates a new
client for each iteration of the retry loop.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>