MGMT-13974: Remove `--wipe` option of `create lso` (#826)
This patch removes the `--wipe` option of the `create lso` command.
Instead of requiring that option the command will write a `ztpfw/wiped:
true|false` annotation in the namespace of the cluster. If the
annotation exists it will not do the wiping. If the annotation
doesn't exist it will do the wiping and add the annotation. This makes
the command idempotent.

Related: https://issues.redhat.com/browse/MGMT-13974

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>