NO-ISSUE: Revert remove deploy workers (#841)
In pull request #162 we removed the step to deploy worker nodes from the
pipeline because with OpenShift 4.12 it is no longer necessary: the
worker nodes can be created from the beginning. But the fixes needed to
support that were done only to the CLI, which was since then removed
from the pipeline. This patch reverts those changes to the pipeline.

Related: https://github.com/rh-ecosystem-edge/ztp-pipeline-relocatable/pull/162

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>